backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.3.4
python-dotenv==1.0.0
pydantic==2.5.3
python-multipart==0.0.6
httpx==0.26.0
pyjwt==2.8.0
celery>=5.3.6
redis>=5.0.1
flower>=2.0.1
polars>=0.20.0
edge-tts==6.1.15
mutagen==1.47.0
imageio-ffmpeg==0.5.1
